+++ bold-pagos-en-linea/assets/js/bold-checkout-ui.js
@@ -1,143 +1,117 @@
 class BoldCheckoutElement extends HTMLElement {
   static get observedAttributes() {
-    return ["plugin_url", "test_mode"];
+    return ["plugin_url", "test_mode", "is_light"];
   }
 
   constructor() {
     super();
     this.shadowDOM = this.attachShadow({ mode: "closed" });
+
+    this.templateHTML = "";
+    this.cssText = "";
+
+    this.loadTemplateAndCSS();
   }
 
-  connectedCallback() {
-    this.render();
-    this.initComponent();
-  }
-
-  attributeChangedCallback(name) {
-    if (name === "plugin_url" || name === "test_mode") {
-      this.render();
-    }
-  }
-  getIcons(pluginUrl, isLight) {
-    const iconsDefault = [
-      `${pluginUrl}../assets/img/payments-method/amex.png`,
-      `${pluginUrl}../assets/img/payments-method/diners.png`,
-      `${pluginUrl}../assets/img/payments-method/discover.png`,
-    ];
-    if (isLight) {
-      return [
-        `${pluginUrl}../assets/img/payments-method/pse_light.png`,
-        `${pluginUrl}../assets/img/payments-method/visa_light.png`,
-        `${pluginUrl}../assets/img/payments-method/mastercard_light.png`,
-        ...iconsDefault,
-        `${pluginUrl}../assets/img/payments-method/bancolombia_light.png`,
-        `${pluginUrl}../assets/img/payments-method/nequi_light.png`,
-      ];
-    }
-    return [
-      `${pluginUrl}../assets/img/payments-method/pse.png`,
-      `${pluginUrl}../assets/img/payments-method/visa.png`,
-      `${pluginUrl}../assets/img/payments-method/mastercard.png`,
-      ...iconsDefault,
-      `${pluginUrl}../assets/img/payments-method/bancolombia.png`,
-      `${pluginUrl}../assets/img/payments-method/nequi.png`,
-    ];
-  }
-
-  async render() {
+  async loadTemplateAndCSS() {
+    const pluginUrl = this.getAttribute("plugin_url") || "";
     try {
-      const pluginUrl = this.getAttribute("plugin_url") || "";
-      const testMode = ["yes", "1", true, 1, "true"].includes(
-        this.getAttribute("test_mode")
-      );
+      const cssResponse = await fetch(`${pluginUrl}../assets/css/bold-checkout-ui.css`);
 
-      const [templateResponse, cssResponse] = await Promise.all([
-        fetch(`${pluginUrl}../assets/html/bold-checkout-ui.html`),
-        fetch(`${pluginUrl}../assets/css/bold-checkout-ui.css`),
-      ]);
+      this.cssText = await cssResponse.text();
 
-      const [templateHTML, cssText] = await Promise.all([
-        templateResponse.text(),
-        cssResponse.text(),
-      ]);
-      const parser = new DOMParser();
-      const doc = parser.parseFromString(templateHTML, "text/html");
-      const modifyTemplate = this.modifyTemplate(doc, templateHTML, testMode);
-
-      const isLight = ["yes", "1", true, 1, "true"].includes(
-        this.getAttribute("is_light")
-      );
-
-      const finalModifyTemplate = this.updateThemeVariable(
-        doc,
-        cssText,
-        modifyTemplate,
-        pluginUrl,
-        isLight
-      );
-
-      const filledTemplate = this.fillTemplate(finalModifyTemplate, pluginUrl);
-
-      this.shadowDOM.innerHTML = `
-          <style>${cssText}</style>
-          ${filledTemplate}
-        `;
+      this.render();
     } catch (error) {
       console.error("Error loading templates:", error);
     }
   }
 
-  fillTemplate(templateHTML, pluginUrl) {
-    return templateHTML.replace(/\{\{plugin_url\}\}/g, pluginUrl);
-  }
-
-  updateThemeVariable(doc, cssText, templateHTML, pluginUrl, isLight) {
-    const elementToRemove = doc.querySelector(
-      "#bold_co_checkout_page_body_payments_method"
-    );
-
-    const icons = this.getIcons(pluginUrl, isLight);
-    icons.forEach((iconUrl) => {
-      const imgElement = doc.createElement("img");
-      imgElement.setAttribute("src", iconUrl);
-      imgElement.setAttribute("alt", "icon");
-      elementToRemove.appendChild(imgElement);
-    });
-
-    this.updateCSSVariable(doc, isLight);
-    return doc.body.innerHTML;
-  }
-
-  modifyTemplate(doc, templateHTML, testMode) {
-    if (!testMode) {
-      const elementToRemove = doc.querySelector(
-        "#bold_co_checkout_page_body_test_mode"
-      );
-
-      if (elementToRemove) {
-        elementToRemove.remove();
-      }
-
-      return doc.body.innerHTML;
+  connectedCallback() {
+    if (this.cssText) {
+      this.render();
     }
-
-    return templateHTML;
   }
 
-  updateCSSVariable(doc, isLight) {
-    if (!isLight) {
+  attributeChangedCallback(name) {
+    if (this.constructor.observedAttributes.includes(name)) {
+      this.render();
+    }
+  }
+
+  render() {
+    if (!this.cssText) {
       return;
     }
-    const element = doc.getElementById("bold_co_checkout_page");
-    element.style.color = "#ffffff";
+
+    this.templateHTML = this.innerHTML;
+
+    const pluginUrl = this.getAttribute("plugin_url") || "";
+    const testMode = ["yes", "1", true, 1, "true"].includes(this.getAttribute("test_mode"));
+    const isLight = ["yes", "1", true, 1, "true"].includes(this.getAttribute("is_light"));
+
+    const filledTemplate = this.fillTemplate(this.templateHTML, pluginUrl, isLight, testMode);
+    this.shadowDOM.innerHTML = `<style>${this.cssText}</style>${filledTemplate}`;
   }
 
-  initComponent() {
-    this.$tag = this.shadowDOM.querySelector(".bold_checkout_element");
+  fillTemplate(templateHTML, pluginUrl, isLight, testMode) {
+    let processedTemplate = templateHTML
+      .replace(/\{\{plugin_url\}\}/g, pluginUrl);
+
+    const parser = new DOMParser();
+    const doc = parser.parseFromString(processedTemplate, "text/html");
+
+    if (!testMode) {
+      const testModeElement = doc.getElementById("bold_co_checkout_page_body_test_mode");
+      if (testModeElement) {
+        testModeElement.remove();
+      }
+    }
+
+    const containerBackgroundElement = doc.getElementById("bold_co_container_info_checkout_page");
+    if(isLight){
+      containerBackgroundElement.classList.add("is_light");
+    }else{
+      containerBackgroundElement.classList.remove("is_light");
+    }
+
+    const iconsUrl = `${pluginUrl}../assets/img/payments-method/`;
+    const icons = this.getIcons(iconsUrl, false);
+    const iconsHTML = icons.map(iconUrl => {
+      const altIcon = iconUrl.replace(iconsUrl, '').split('.')[0];
+      return `<img src="${iconUrl}" alt="${altIcon}" />`;
+    }).join("");
+
+    const iconsElement = doc.getElementById("bold_co_checkout_page_body_payments_method");
+    iconsElement.innerHTML = iconsHTML;
+
+    processedTemplate = doc.body.innerHTML;
+
+    return processedTemplate;
   }
 
-  disconnectedCallback() {
-    this.remove();
+  getIcons(iconsUrl, isLight) {
+    const iconsDefault = [
+      `${iconsUrl}amex.png`,
+      `${iconsUrl}diners.png`,
+      `${iconsUrl}discover.png`,
+    ];
+    return isLight
+      ? [
+          `${iconsUrl}pse_light.png`,
+          `${iconsUrl}visa_light.png`,
+          `${iconsUrl}mastercard_light.png`,
+          ...iconsDefault,
+          `${iconsUrl}bancolombia_light.png`,
+          `${iconsUrl}nequi_light.png`,
+        ]
+      : [
+          `${iconsUrl}pse.png`,
+          `${iconsUrl}visa.png`,
+          `${iconsUrl}mastercard.png`,
+          ...iconsDefault,
+          `${iconsUrl}bancolombia.png`,
+          `${iconsUrl}nequi.png`,
+        ];
   }
 }