+++ bold-pagos-en-linea/bold-co.php
@@ -3,13 +3,13 @@
  * Plugin Name: Bold pagos en linea 
  * Plugin URI: https://developers.bold.co/pagos-en-linea/boton-de-pagos/plugins/wordpress
  * Description: Recibe pagos en tu tienda de forma segura con los métodos de pago más usados y con la mejor experiencia para tus clientes.
- * Version: 3.0.6
+ * Version: 3.1.0
  * Author: Bold
  * Author URI: http://www.bold.co/
  * Network: true
  * Text Domain: bold-pagos-en-linea
  * WC requires at least: 5.5.2
- * WC tested up to: 8.1.0
+ * WC tested up to: 9.4.1
  * Requires PHP: 7.4
  * License: GPLv3
  * License URI: https://www.gnu.org/licenses/gpl-3.0.html
@@ -60,7 +60,7 @@ use BoldPagosEnLinea\BoldConstants;
 
 // Función para registrar y cargar el script de botón de pago
 function bold_co_custom_header_code(): void {
-    wp_register_script('woocommerce_bold_payment_button_js', BoldConstants::URL_CHECKOUT.'/library/boldPaymentButton.js', [], '3.0.6', true);
+    wp_register_script('woocommerce_bold_payment_button_js', BoldConstants::URL_CHECKOUT.'/library/boldPaymentButton.js', [], '3.1.0', true);
     wp_enqueue_script('woocommerce_bold_payment_button_js');
 }